测试Roboduino电机控制程序, 测试基本通过

+++ hcr/arduino/my_libs/Roboduino/examples/TryMotor1/TryMotor1.pde
@@ -34,8 +34,8 @@ void loop()
     // Êä³öËÙ¶È
     
     {
-        uint8_t m1 = RoboduinoMotor.motorRead(0);
-        uint8_t m2 = RoboduinoMotor.motorRead(1);
+        int m1 = RoboduinoMotor.motorRead(0);
+        int m2 = RoboduinoMotor.motorRead(1);
         
         Serial.print("M1: ");
         Serial.println(m1);
